+++ backend/aci/common/logging_setup.py
@@ -3,14 +3,6 @@ from logging.handlers import RotatingFileHandler
 
 import logfire
 
-from aci.server.context import (
-    agent_id_ctx_var,
-    api_key_id_ctx_var,
-    org_id_ctx_var,
-    project_id_ctx_var,
-    request_id_ctx_var,
-)
-
 
 # the setup is called once at the start of the app
 def setup_logging(
@@ -59,29 +51,3 @@ def get_logger(name: str, level: int = logging.INFO) -> logging.Logger:
     logger = logging.getLogger(name)
     logger.setLevel(level)
     return logger
-
-
-class RequestContextFilter(logging.Filter):
-    def filter(self, record: logging.LogRecord) -> bool:
-        # Only add attributes when values are not None
-        request_id = request_id_ctx_var.get()
-        if request_id is not None:
-            record.__dict__["request_id"] = request_id
-
-        api_key_id = api_key_id_ctx_var.get()
-        if api_key_id is not None:
-            record.__dict__["api_key_id"] = api_key_id
-
-        project_id = project_id_ctx_var.get()
-        if project_id is not None:
-            record.__dict__["project_id"] = project_id
-
-        agent_id = agent_id_ctx_var.get()
-        if agent_id is not None:
-            record.__dict__["agent_id"] = agent_id
-
-        org_id = org_id_ctx_var.get()
-        if org_id is not None:
-            record.__dict__["org_id"] = org_id
-
-        return True
